Update to Couchbase Lite 1.0
CBL 1.0 is not on Cocoapods, therefore added the framework as dependency.
This seemed to miss some CBL dependencies, in order to get it working, I
also had to add FBDB pod and a weak link to /usr/lib/libz.dylib.
Probably really hacky, need to find a better way for the dependencies